MSVC fix

+++ include/hpp_proto/timestamp_codec.hpp
@@ -13,9 +13,9 @@ struct timestamp_codec {
 
     assert(val >= 0);
     if constexpr (Len == 9) {
-      const uint64_t hi = val / 100000000;
+      const int hi = val / 100000000;
       assert(hi < 10);
-      *buf++ = '0' + hi;
+      *buf++ = '0' + static_cast<uint8_t>(hi);
       buf = glz::to_chars_u64_len_8(buf, uint32_t(val % 100000000));
     } else if constexpr (Len == 4) {
       buf = glz::to_chars_u64_len_4(buf, uint32_t(val));
